Move hex conversion functions to top of crate
Moves the hex conversion functions to the top-level of the crate, since
they may be used throughout the library.

Also adds a new `HexError` type that is used for the `hex_str_to_bytes`
function, since the older `Error` type the function used is no longer
appropriate with its new location.

Signed-off-by: Logan Seeley <seeley@bitwise.io>