Move secp256k1 tests to test crate
Moves and adapts the tests in the `cylinder::secp256k1` module to be
implementation-agnostic in the test crate. This will allow these tests
to be used across implementation as needed.

The tests that are moved have been simplified where possible.

The following tests have been added to replace the old secp256k1 tests:
- `get_public_key` and `signer_public_key` replace the old
  `priv_to_public_key` test
- `signing` and `signing_and_verification` replace the old
  `single_key_signing` test
- `verification` and `signing_and_verification` replace the old
  `verification` test
- `verification_failure` replaces the old `verification_error` test
- `multiple_key_signing_and_verification` replaces the old
  `many_key_signing` test
- `signer_return_from_fn` replaces the old
  `single_key_signing_return_from_func` test

Signed-off-by: Logan Seeley <seeley@bitwise.io>