Implement JsonWebToken Builder and Parser
This commit implements building and parsing of the custom cylinder JSON
web token values.

The process is divided between the JsonWebTokenBuilder, which is used to
construct encoded, signed token.  As of this commit, the header values
and claims values are a flat object of (String, String) pairs.

The token is read and verified by the JsonWebTokenParser.  This parser
wraps a Verifier instance which provides the signature verification.  It
produces JsonWebToken struct that can be used to examine the token's
claims.

As of this commit, the algorithm choice for the JWT (the field "alg" in
the header) is hard-coded to "secp256k1", as this is currently the only
signing algorithm implemented in Cylinder.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>