Add cyl CLI with "jwt generate" subcommand
This change introduces a `cyl` CLI tool with an initial "jwt generate"
subcommand.  This command uses a key from the usual places and generates
a basic cylinder JWT.  This JWT can be used in conjunction with
commands, such as using curl against a REST API with a Cylinder JWT
authorization token.

It also adds the cli directory to the list of crates in the justfile.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>